\name{inverse.kosambi}
\alias{inverse.kosambi}
\title{inverse Kosambi}
\usage{
  inverse.kosambi(x)
}
\arguments{
  \item{x}{a numeric vector of distance values, all values
  in range [0, 0.5]}
}
\value{
  a numeric vector of inverse Kosambi distances
}
\description{
  Calculate the mab distance using inverse Kosambi
  function. NB, x must be in range [0, 0.5]
}
\author{
  Mark Cowley, 30 June 2005
}
\references{
  \url{http://www.mapmanager.org/MMM/MMMAlgorithms.html#996839}
}

