\name{normrnd}
\alias{normrnd}
\title{create a matrix of random values}
\usage{
  normrnd(mean = 0, sd = 1)
}
\arguments{
  \item{mean}{a vector or matrix of mean values}

  \item{sd}{a vector or matrix of sd values}
}
\value{
  a vector or matrix of normally-distributed, random
  variables
}
\description{
  Matlab offers the \code{normrnd} function which allows
  more control over creating matrices of random variables,
  allowing matrices of \code{mean} and \code{sd} input
  parameters.
}
\author{
  Mark Cowley, 2012-07-16
}

