\name{standardise}
\alias{standardise}
\title{Standardise}
\usage{
  standardise(x)
}
\arguments{
  \item{x}{a numeric vector}
}
\value{
  a numeric vector
}
\description{
  Standardise a vector of numbers, by subtracting the mean,
  and dividing by the stdev
}
\author{
  Mark Cowley, 2008-06-11
}

